requirements.txt
pyqt
pycryptodome
